GAN/requirements.txt
Pillow==11.1.0
opencv-python-headless==4.11.0.86
numpy==1.26.4
torch==2.6.0
torchvision==0.21.0
open-clip-torch==2.30.0
datasets==3.2.0
scikit-image==0.25.1
scipy==1.15.2
python-dotenv==1.0.1
tqdm==4.67.1
scikit-learn==1.6.1
